feat: Enhance documentation and add new modules
Expanded the documentation with descriptions and settings details for several new combat modules, including BurstClicker, Criticals, HitSelect, HitBox, InfiniteAura, JumpReset, KillAura, and MoreKB. This provides users with comprehensive information to understand and utilize these modules effectively.

Additionally, the documentation was restructured and revamped with Tailwind CSS for improved readability and a more visually appealing presentation. Reviews and discord link was also added to the home page.